Fixed problems linked to merging
