+++ app/assets/stylesheets/components/_container.scss
@@ -1,4 +1,4 @@
-.container {
+.container-change-name {
   margin-top: 106px;
   margin-bottom: 95px;
 }
