+++ app/javascript/packs/application.js
@@ -24,32 +24,22 @@ require("channels")
 
 // External imports
 import "bootstrap";
-import { initStarRating } from '../plugins/init_star_rating';
-import Glide from "@glidejs/glide";
 
 // Internal imports, e.g:
 // import { initSelect2 } from '../components/init_select2';
-import { initStarRating } from '../plugins/init_star_rating';
+// import { initStarRating } from '../plugins/init_star_rating';
 import { initMapbox } from '../plugins/init_mapbox';
+import { initGlide } from "../plugins/init_glide";
 
 document.addEventListener('turbolinks:load', () => {
   // Call your functions here, e.g:
   // initSelect2();
 
   // carousels
-  const config_1 = {
-    type : 'carousel',
-  }
-  const config_2 = {
-    type : 'carousel',
-    perView: 2
-  }
-  new Glide('.glide-trendy-bars', config_1).mount()
-  new Glide('.glide-events-nearby', config_2).mount()
-  new Glide('.glide-based-profile', config_2).mount()
+  initGlide();
 
   // stars rating
-  initStarRating();
+  //initStarRating();
   
   //mapbox
   initMapbox();
